+++ __init__.py
@@ -1,1 +1,6 @@
-"""astrbot_plugin_mimo - 小米 MiMo 平台用量查询 AstrBot 插件"""
+"""astrbot_plugin_resource_query - 多平台资源查询 AstrBot 插件
+
+支持平台：
+  - MiMo：小米 MiMo 平台用量查询
+  - 华数广电：流量/通话/余额查询
+"""